fix unsigned long bug

+++ src/cpp/Commands/Double.cpp
@@ -15,8 +15,8 @@ Double::Double() {
 
 // Execute method
 void Double::executeCommand(CandyGameState& gameState) {
-    unsigned long long oldPoints = gameState.getPointPool();
-    unsigned long long newPoints = oldPoints * (unsigned long long)2;
+    uint64_t oldPoints = gameState.getPointPool();
+    uint64_t newPoints = oldPoints * (uint64_t)2;
     gameState.setPointPool(newPoints);
 
     cout << gameState.getCurrentTurnPlayer().getName() << " melakukan DOUBLE! Poin hadiah naik dari " << oldPoints << " menjadi " << newPoints << "!" << endl;
